refactor static info
