+++ aptos-move-witnesses/Cargo.toml
@@ -7,6 +7,10 @@ edition = "2021"
 move-vm-runtime = { workspace = true }
 move-core-types = { workspace = true }
 move-vm-types = { workspace = true }
+move-package = { workspace = true }
+move-binary-format = { workspace = true }
+logger = { path = "../common/logger" }
+movelang = { path = "../movelang" }
 types = { path = "../common/types" }
 serde = { version = "1.0.203", features = ["derive"] }
 strum_macros = { workspace = true }
