+++ vm-circuit/src/chips/execution_chip_v2/lookup_table/byecode_table.rs
@@ -1,6 +1,7 @@
 use crate::chips::execution_chip_v2::lookup_table::utils::assign_fixed_table;
+use crate::chips::execution_chip_v2::lookup_table::utils::ToFields;
 use crate::table::LookupTable;
-use crate::witness::static_info::StaticInfo;
+use aptos_move_witnesses::static_info::StaticInfo;
 use halo2_proofs::circuit::Layouter;
 use halo2_proofs::plonk::{Any, Column, ConstraintSystem, Error, Fixed};
 use types::Field;
@@ -46,7 +47,7 @@ impl BytecodeLookupTable {
         let field_elements: Vec<Vec<F>> = static_info
             .bytecode_info
             .iter()
-            .map(|row| row.to_fe())
+            .map(|row| row.to_fields())
             .collect();
         assign_fixed_table(layouter, self.columns(), &field_elements, "bytecode_table")
     }
