+++ vm-circuit/src/chips/execution_chip_v2/lookup_table/constant_table.rs
@@ -1,13 +1,11 @@
 // Copyright (c) zkMove Authors
 
-use crate::chips::execution_chip_v2::lookup_table::constant_table::flatten::Flatten;
 use crate::chips::execution_chip_v2::lookup_table::utils::assign_fixed_table;
+use crate::chips::execution_chip_v2::lookup_table::utils::ToFields;
 use crate::chips::execution_chip_v2::step_v2::NUM_OF_VALUE_LIMBS;
-use crate::chips::execution_chip_v2::utils::to_field::ToField;
 use crate::table::LookupTable;
-use crate::witness::static_info::constant::ConstantInfo;
-use crate::witness::static_info::StaticInfo;
-use aptos_move_witnesses::utils::SubIndexUtils;
+use aptos_move_witnesses::static_info::constant::flatten::Flatten;
+use aptos_move_witnesses::static_info::StaticInfo;
 use aptos_move_witnesses::ValueItem;
 use halo2_proofs::circuit::Layouter;
 use halo2_proofs::plonk::{Any, Column, ConstraintSystem, Error, Fixed};
@@ -48,8 +46,18 @@ impl ConstantLookupTable {
             .constant_info
             .iter()
             .flat_map(|c| {
-                let rows: Vec<ConstantTableRow> = c.clone().into();
-                rows.iter().map(|row| row.to_fe()).collect::<Vec<_>>()
+                let rows: Vec<_> = c
+                    .value
+                    .clone()
+                    .flatten(vec![0])
+                    .iter()
+                    .map(|item| ConstantTableRow {
+                        module_index: c.module_index,
+                        constant_index: c.constant_index,
+                        value_item: item.clone(),
+                    })
+                    .collect::<Vec<_>>();
+                rows.iter().map(|row| row.to_fields()).collect::<Vec<_>>()
             })
             .collect();
         assign_fixed_table(layouter, self.columns(), &field_elements, "constant_table")
@@ -77,104 +85,7 @@ impl<F: Field> LookupTable<F> for ConstantLookupTable {
 
 #[derive(Clone, Eq, PartialEq, Debug)]
 pub struct ConstantTableRow {
-    module_index: usize,
-    constant_index: usize,
-    value_item: ValueItem,
-}
-
-impl ConstantTableRow {
-    fn to_fe<F: Field>(&self) -> Vec<F> {
-        vec![
-            F::from_u128(self.module_index as u128),
-            F::from_u128(self.constant_index as u128),
-            F::from_u128(self.value_item.sub_index.into_u128()),
-        ]
-        .into_iter()
-        .chain(self.value_item.value.to_field())
-        .chain(vec![F::from(self.value_item.header as u64)])
-        .collect()
-    }
-}
-impl From<ConstantInfo> for Vec<ConstantTableRow> {
-    fn from(constant: ConstantInfo) -> Vec<ConstantTableRow> {
-        let items = constant.value.flatten(vec![0]);
-        items
-            .iter()
-            .map(|item| ConstantTableRow {
-                module_index: constant.module_index,
-                constant_index: constant.constant_index,
-                value_item: item.clone(),
-            })
-            .collect()
-    }
-}
-
-pub mod flatten {
-    use aptos_move_witnesses::step_state::SubIndex;
-    use aptos_move_witnesses::sub_index;
-    use aptos_move_witnesses::utils::ValueHeader;
-    use aptos_move_witnesses::{SimpleValue, ValueItem};
-    use move_core_types::value::MoveValue;
-
-    pub trait Flatten {
-        fn flatten(self, sub_index: SubIndex) -> Vec<ValueItem>;
-        fn flen(&self) -> usize;
-    }
-
-    impl Flatten for MoveValue {
-        fn flatten(self, sub_index: SubIndex) -> Vec<ValueItem> {
-            let flen = self.flen();
-            match self {
-                MoveValue::U8(u) => vec![value_item(sub_index, SimpleValue::U8(u))],
-                MoveValue::U16(u) => vec![value_item(sub_index, SimpleValue::U16(u))],
-                MoveValue::U32(u) => vec![value_item(sub_index, SimpleValue::U32(u))],
-                MoveValue::U64(u) => vec![value_item(sub_index, SimpleValue::U64(u))],
-                MoveValue::U128(u) => vec![value_item(sub_index, SimpleValue::U128(u))],
-                MoveValue::U256(u) => vec![value_item(sub_index, SimpleValue::U256(u))],
-                MoveValue::Bool(b) => vec![value_item(sub_index, SimpleValue::Bool(b))],
-                MoveValue::Vector(values) => {
-                    let len = values.len();
-                    let mut items = Vec::new();
-                    items.push(header_item(sub_index.clone(), flen, len));
-
-                    for (i, value) in values.into_iter().enumerate() {
-                        let value_sub_index = sub_index::concat(sub_index.clone(), vec![i + 1]);
-                        let mut flattened_value = value.flatten(value_sub_index);
-                        items.append(&mut flattened_value);
-                    }
-                    items
-                }
-                _ => unimplemented!(),
-            }
-        }
-
-        fn flen(&self) -> usize {
-            match self {
-                MoveValue::U8(_)
-                | MoveValue::U16(_)
-                | MoveValue::U32(_)
-                | MoveValue::U64(_)
-                | MoveValue::U128(_)
-                | MoveValue::U256(_)
-                | MoveValue::Bool(_) => 1,
-                MoveValue::Vector(values) => values.iter().fold(0, |sum, v| sum + v.flen()) + 1,
-                _ => unimplemented!(),
-            }
-        }
-    }
-
-    fn value_item(sub_index: SubIndex, simple: SimpleValue) -> ValueItem {
-        ValueItem {
-            sub_index,
-            header: false,
-            value: simple,
-        }
-    }
-    fn header_item(sub_index: SubIndex, flen: usize, len: usize) -> ValueItem {
-        ValueItem {
-            sub_index,
-            header: true,
-            value: ValueHeader::new(flen as u16, len as u16).into(),
-        }
-    }
+    pub module_index: usize,
+    pub constant_index: usize,
+    pub value_item: ValueItem,
 }
